feat: split exchange section into two sections
