+++ src/sections/StrenghtsSection.tsx
@@ -1,0 +1,7 @@
+type Props = {
+  strengths: Queries.IndexPageQuery['strengths'];
+};
+
+export function StrengthsSection({ strengths }: Props) {
+  return <article />;
+}
